filiales_django/requirements/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==24.3
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
faker==25.0.0
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==4.1.0
black==24.8.0
ruff==0.5.5
isort==5.13.2
